Move `SumHint` and `PathHint`
